pyproject.toml
[project]
name = "dataforge"
version = "0.1.0"
description = "Scaffold for a CLI-first data-quality repair agent."
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11,<3.13"
keywords = ["data-quality", "ai-agent", "llm", "rl", "smt", "dbt"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pandas>=2.2",
    "pyarrow>=16.0",
    "pydantic>=2.7",
    "typer>=0.12",
    "rich>=13.7",
    "z3-solver>=4.13",
    "networkx>=3.3",
    "causal-learn>=0.1.4",
    "scipy>=1.13",
    "httpx>=0.27",
    "tenacity>=8.3",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "sqlglot>=25.0",
    "duckdb>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-benchmark>=4.0",
    "pytest-xdist>=3.6",
    "hypothesis>=6.100",
    "mutmut>=3.5",
    "ruff>=0.11",
    "mypy>=1.10",
    "pandas-stubs>=2.2",
    "types-PyYAML",
]
train = [
    "transformers>=4.44",
    "datasets>=2.20",
    "trl>=1.0",
    "accelerate>=0.33",
    "peft>=0.12",
    "bitsandbytes>=0.43",
]
eval = [
    "matplotlib>=3.9",
    "seaborn>=0.13",
]
playground = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
]
all = [
    "dataforge[dev,train,eval,playground]",
]

[project.scripts]
dataforge = "dataforge.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["dataforge*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "PIE", "RET", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
strict = true
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
explicit_package_bases = true
exclude = [
    "^[^/]*\\.py$",          # loose root-level scripts (hackathon legacy)
    "^(server|training|playground|benchmark_results|datasets)/",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --strict-config"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "requires_network: tests that need internet access",
    "requires_llm: tests that call a free-tier LLM API",
]

[tool.coverage.run]
source = ["dataforge"]
branch = true

[tool.coverage.report]
fail_under = 90
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
